kamu-datasets-services, DatasetEntryIndexer: extract from DatasetEntryService
